+++ src/controls/baseControl.ts
@@ -276,8 +276,6 @@ export default abstract class WUPBaseControl<
     // NiceToHave: change icons to fonts: wupfont: in this case possible to improve btnClear: hover
   }
 
-  // todo change margin 20px to --base-margin;
-
   /** StyleContent related to component */
   static get $style(): string {
     // WARN: 'contain:style' is tricky rule
@@ -286,6 +284,7 @@ export default abstract class WUPBaseControl<
         --ctrl-focus-border: var(--ctrl-focus);
         contain: style;
         display: block;
+        flex: 1;
         margin: var(--base-margin) 0;
         box-shadow: 0 0 0 1px var(--ctrl-border);
         border-radius: var(--ctrl-border-radius);