+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 ---
 
 - [ModalElement](src/modalElement.ts.ts). _Wrong behavior on confirmModal + replace: true + click outside on previous/hidden modal_
-- **Controls**. _Missed css-var `--base-margin`_
+- **Controls**
+  - _Missed css-var `--base-margin`_
+  - _Missed style `flex: 1` for switch/checkbox_ controls
 
 ---
 
@@ -58,6 +60,7 @@
 
   - Added static `.$use()`. Call it before using element: `WUPTextControl.$use()` (_self-registration by import_ works now and will be removed in the future)
   - Renamed css-vars:
+
     - `--ctrl-select-menu-hover` >>> `--menu-hover-bg`
     - `--anim-time` >>> `--anim-t`
     - `--popup` >>> `popup-text`
@@ -91,7 +94,7 @@
     - enum `HideCases` >>> `MenuHideCases`
     - event `$showMenu` >>> `$openMenu`
     - event `$hideMenu` >>> `$closeMenu`
-    etc.
+      etc.
 
 ---